pyproject.toml
[tool.poetry]
name = "bot_test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.30.0"
aiogram = "^3.20.0.post0"
redis = "^6.2.0"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
asyncio = "^3.4.3"
yoyo-migrations = "^9.0.0"
python-multipart = "^0.0.20"
uvicorn = "^0.34.3"
fastapi = "^0.115.12"
jinja2 = "^3.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
